==> 2_pdm_autism_frequentist_stats.Rmd <==
---
title: "cbs_workflow_v3_accuracy"
author: "Nazia Jassim"
date: "06/11/2020"
output: html_document
---

## R Markdown

```{r}
getwd();
library(tidyverse);
library(car);
library(stats);
library(rstatix);

# for interaction
library(nlme);
library(effects);
library(psych)


```

##FEATURE MATCH

```{r}
#read.csv("feature_match_accuracy_rate.csv") %>% 
read.csv("fm_accuracy.csv") %>%
    mutate(cohort=factor(cohort, levels= c("asc", "ctr"), labels=c("Autism", "Control")),
          condition= factor(condition, levels=c("match", "mismatch"),labels=c("Match", "Mismatch")), 
         Sex = factor(sex)) -> feature_match_accuracy;


fm_boxplot <-ggplot(feature_match_accuracy, mapping= aes(x=cohort, y= accuracy_rate_per_condition, fill=cohort)) +
  xlab("Group")+ ylab("Accuracy Rate")+ 
  labs(fill= "Group", title = , subtitle = "Task 2: Feature Match")+
  facet_grid(~condition) +
    #ggdist::stat_halfeye(adjust = .5, width = .7, .width = 0, justification = -.2, point_colour = NA) + 
    geom_boxplot(width = .6, outlier.shape = NA) + 
    geom_jitter(width = 0.2, size= 1, alpha= 0.3) 

fm_boxplot+ scale_fill_manual(values=c("#96c5bd","#C5969E")) + theme_minimal() + 
    theme(plot.title = element_text(size= 10),  plot.subtitle = element_text(size = 12, face = "bold", hjust = 0.5))

ggsave("fm_boxplot.png", width=6, height=3)

########
## boxplots with Interaction lines
ggplot(feature_match_accuracy, mapping= aes(x=condition, y= accuracy_rate_per_condition, fill=condition)) +
    facet_wrap(~cohort, ncol=2) + 
    geom_boxplot(width = .5, outlier.shape = NA) + 
    geom_jitter(width = .09, size= 1, alpha= 0.3) + 
    geom_line(aes(y= accuracy_rate_per_condition, group = condition), col = "darkgrey", size = 0.40, alpha=0.4)



# female only--------------------------------------
read.csv("fm_accuracy.csv") %>%
    mutate(cohort=factor(cohort, levels= c("asc", "ctr"), labels=c("Autism", "Control")),
          condition= factor(condition, levels=c("match", "mismatch"),labels=c("Match", "Mismatch")), 
         Sex = factor(sex)) %>% filter(sex=="Female")  -> feature_match_accuracy_female

feature_match_accuracy_female %>% group_by(cohort) %>% summarise(mean= mean(accuracy_rate_per_condition), sd= sd(accuracy_rate_per_condition))

#-------------------------------------------

#plot 
fm_boxplot <-ggplot(feature_match_accuracy, mapping= aes(x=cohort, y= accuracy_rate_per_condition, fill=cohort)) +
  xlab("Group")+ ylab("Accuracy Rate")+ 
  labs(fill= "Group", title = "Task 2: Feature Match")+
  stat_boxplot(geom='errorbar', linetype=1, width=0.1)+ 
  geom_boxplot(outlier.shape=1)+ 
  geom_jitter(color="black", size=0.4, alpha=0.9) + 
  facet_wrap(~condition, ncol=2) 

fm_boxplot+ scale_fill_manual(values=c('#f1a340','#998ec3')) + theme_minimal()

#ggplot(feature_match_accuracy, mapping= aes(x=condition, y= accuracy_rate_per_condition, fill=cohort))+ geom_boxplot()+ geom_jitter(color="black", size=0.4, alpha=0.9)+ facet_wrap(~sex, ncol=2)

## CONTRASTS

## Orthogonal contrasts
###1.  Cohort- control -1, ASC 1 2) Condition- Match 1 , Mismatch -1
contrasts(feature_match_accuracy$cohort)<- c(-1,1)
contrasts(feature_match_accuracy$condition)<- c(1,-1)


##ANOVA

# model 1- 2x2 factorial anova 
feature_match_accuracy_model1 <- lm(accuracy_rate_per_condition ~ cohort + condition + cohort:condition, data = feature_match_accuracy);

Anova(feature_match_accuracy_model1, type= "III")


## model 2 - control for age

feature_match_accuracy_model2 <- lm(accuracy_rate_per_condition ~ cohort + condition + cohort:condition+ Age, data = feature_match_accuracy);

Anova(feature_match_accuracy_model2, type= "III")


# model 3- control for sex
feature_match_accuracy_model3 <- lm(accuracy_rate_per_condition ~ cohort + condition + cohort:condition+ sex, data = feature_match_accuracy);

Anova(feature_match_accuracy_model3, type= "III")


## ANOVA Assumptions
plot(feature_match_accuracy_model1, which = 2, add.smooth = FALSE)

## Normality
feature_match_accuracy%>%
    group_by(cohort, condition) %>%
    shapiro_test(accuracy_rate_per_condition)

## Homogeneity of variance, for each factor
leveneTest(accuracy_rate_per_condition~cohort, data= feature_match_accuracy);
leveneTest(accuracy_rate_per_condition~condition, data=feature_match_accuracy)
leveneTest(accuracy_rate_per_condition~age, data=feature_match_accuracy)
  

## Interaction plot with error bar
fm_accuracy_means <- 
    feature_match_accuracy %>% 
    group_by(cohort, condition) %>% # <- remember to group by *both* factors
    summarise(Mean_Accuracy_Rate = mean(accuracy_rate_per_condition),
              sem = (sd(accuracy_rate_per_condition)/sqrt(length(accuracy_rate_per_condition))))

fm_interaction <- ggplot(fm_accuracy_means, 
       aes(x = condition, y = Mean_Accuracy_Rate, color = cohort)) + geom_line(aes(group=cohort), size=1)+ geom_linerange(aes(x= condition, ymin = Mean_Accuracy_Rate - sem, ymax = Mean_Accuracy_Rate + sem), size=0.5)+
  geom_point(size=3, shape= "square") + 
  scale_color_manual(values= c('#f1a340','#998ec3'))+ 
  labs(
         title = "B",  
         subtitle = "Experiment 2: Feature Match",
           x = "Condition",
           y = "Accuracy Rate", 
           color= "Group");

fm_interaction + theme_minimal() + 
    theme(plot.title = element_text(size= 10),  plot.subtitle = element_text(size = 12, face = "bold", hjust = 0.5))
## POST HOC - 

#library(compute.es); library(multcomp)
```




##POLYGONS
```{r}
read.csv("polygons_accuracy_rate.csv") %>% mutate(cohort=factor(cohort, levels= c("asc", "ctr"), labels=c("Autism", "Control")),
          condition= factor(condition, levels=c("match", "mismatch"),labels=c("Match", "Mismatch")), 
         Sex = factor(sex))-> polygons_accuracy;


#### --- female only -----------------------------------------
read.csv("polygons_accuracy_rate.csv") %>% mutate(cohort=factor(cohort, levels= c("asc", "ctr"), labels=c("Autism", "Control")),
          condition= factor(condition, levels=c("match", "mismatch"),labels=c("Match", "Mismatch")), 
         Sex = factor(sex)) %>% filter(sex== "Female")-> polygons_accuracy_female;

polygons_accuracy_female %>% group_by(cohort, condition) %>% summarise(mean = mean(accuracy_overall), sd= sd(accuracy_rate_per_condition
                                                                             ))

### ---------------------------------------

#plot

poly_boxplot <-ggplot(polygons_accuracy, mapping= aes(x=cohort, y= accuracy_rate_per_condition, fill=cohort)) +
  xlab("Group")+ ylab("Accuracy Rate")+ 
  labs(fill= "Group", title = , subtitle = "Task 1: Interlocking Polygons")+
  facet_grid(~condition) +
    #ggdist::stat_halfeye(adjust = .5, width = .7, .width = 0, justification = -.2, point_colour = NA) + 
     geom_boxplot(width = .6, outlier.shape = NA) + 
    geom_jitter(width = 0.2, size= 1, alpha= 0.3) 

poly_boxplot+ scale_fill_manual(values=c("#96c5bd","#C5969E")) + theme_minimal() + 
    theme(plot.title = element_text(size= 10),  plot.subtitle = element_text(size = 12, face = "bold", hjust = 0.5))


ggsave("poly_boxplot.png", width=6, height=3)


fm_boxplot <-ggplot(feature_match_accuracy, mapping= aes(x=cohort, y= accuracy_rate_per_condition, fill=cohort)) +
  xlab("Group")+ ylab("Accuracy Rate")+ 
  labs(fill= "Group", title = "", subtitle = "Task 2: Feature Match")+
  facet_grid(~condition) +
    #ggdist::stat_halfeye(adjust = .5, width = .7, .width = 0, justification = -.2, point_colour = NA) + 
    geom_boxplot(width = .5, outlier.shape = NA) + 
    geom_jitter(width = .09, size= 1, alpha= 0.3) 

fm_boxplot+ scale_fill_manual(values=c('#f1a340','#998ec3')) + theme_minimal() + 
    theme(plot.title = element_text(size= 10),  plot.subtitle = element_text(size = 12, face = "bold", hjust = 0.5))
## ANOVA Assumptions

## Normality
polygons_accuracy%>%
    group_by(cohort, condition) %>%
    shapiro_test(accuracy_rate_per_condition)

## Homogeneity of variance, for each factor
leveneTest(accuracy_rate_per_condition~cohort, data= polygons_accuracy)
leveneTest(accuracy_rate_per_condition~condition, data= polygons_accuracy)


## CONTRASTS

## Orthogonal contrasts
###1.  Cohort- control -1, ASC 1 2) Condition- Match 1 , Mismatch -1
contrasts(polygons_accuracy$cohort)<- c(-1,1)
contrasts(polygons_accuracy$condition)<- c(1,-1)


##ANOVA

# model 1- 2x2 factorial anova 
polygons_accuracy_model1 <- lm(accuracy_rate_per_condition ~ cohort + condition + cohort:condition, data = polygons_accuracy)

Anova(polygons_accuracy_model1, type= "III")



## Interaction plot
polygons_accuracy_means <- 
    polygons_accuracy %>% 
    group_by(cohort, condition) %>% # <- remember to group by *both* factors
    summarise(Mean_Accuracy_Rate = mean(accuracy_rate_per_condition), sem = sd(accuracy_rate_per_condition)/sqrt(length(accuracy_rate_per_condition)))

ggplot(polygons_accuracy_means, 
       aes(x = cohort, y = Mean_Accuracy_Rate, color = condition)) + geom_line(aes(group=condition), size=1)+ geom_linerange(aes(x= cohort, ymin = Mean_Accuracy_Rate - sem, ymax = Mean_Accuracy_Rate + sem), size=0.3)+ geom_point(size=3) + scale_color_manual(values= c('#f1a340','#998ec3'))+ labs(
           title = "pOLY",
           x = "Condition",
           y = "Accuracy Rate", 
           color= "Condition") + theme_minimal()
 

ggplot(polygons_accuracy_means, 
       aes(x = condition, y = Mean_Accuracy_Rate, colour = cohort, group = cohort)) + geom_point(size = 4) + geom_line()+ scale_colour_manual(values= c('#f1a340','#998ec3'))+ labs(
    title = "B",
    x = "Condition",
    y = "Accuracy Rate", 
    fill= "Group",
    colour= "Group"
)

geom_point(size = 4) + geom_line()+ scale_color_manual(values= c('#f1a340','#998ec3'))+ labs(
           title = "Experiment 2: Feature Match",
           x = "Group",
           y = "Accuracy Rate", 
           fill= "Condition")



```




## ROTATIONS
```{r}
rotations_accuracy <- read.csv("rotations_accuracy_rate.csv")
rotations_accuracy <- mutate(rotations_accuracy, condition= as_factor(condition), cohort=as_factor(cohort), sex = as_factor(sex));

#plot
ggplot(rotations_accuracy, mapping= aes(x=cohort, y= accuracy_rate_per_condition, fill=cohort))+ geom_boxplot()+ geom_jitter(color="black", size=0.4, alpha=0.9) + facet_wrap(~condition, ncol=2)+ scale_fill_manual(values= c("#006767", "#ffb348"))+ labs(
    title = "Rotations Accuracy Rates",
    x = "Group",
    y = "Accuracy Rate", 
    fill= "Group"
)


##
# ANOVA Assumptions

## Normality
rotations_accuracy%>%
    group_by(cohort, condition) %>%
    shapiro_test(accuracy_rate_per_condition)

## Homogeneity of variance, for each factor
leveneTest(accuracy_rate_per_condition~cohort, data= rotations_accuracy)
leveneTest(accuracy_rate_per_condition~condition, data= rotations_accuracy)


## CONTRASTS

## Orthogonal contrasts
###1.  Cohort- control -1, ASC 1 2) Condition- Match 1 , Mismatch -1
contrasts(rotations_accuracy$cohort)<- c(-1,1)
contrasts(rotations_accuracy$condition)<- c(1,-1)


##ANOVA

# model 1- 2x2 factorial anova 
rotations_accuracy_model1 <- lm(accuracy_rate_per_condition ~ cohort + condition + cohort:condition, data = rotations_accuracy)

Anova(rotations_accuracy_model1, type= "III")



## Interaction plot
rotations_accuracy_means <- 
    rotations_accuracy %>% 
    group_by(cohort, condition) %>% # <- remember to group by *both* factors
    summarise(Mean_Accuracy_Rate = mean(accuracy_rate_per_condition))

ggplot(rotations_accuracy_means, 
       aes(x = cohort, y = Mean_Accuracy_Rate, colour = condition, group = condition)) + geom_point(size = 4) + geom_line()+ scale_colour_manual(values= c("#006767", "#ffb348"))+ labs(
    title = "Rotations Accuracy Rates",
    x = "Group",
    y = "Accuracy Rate", 
    fill= "Group"
)

```
